ecom_data_pipeline/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
requests
beautifulsoup4
apache-airflow-providers-google
apache-airflow-providers-snowflake

